Made classes implement Serializable. These Classes include: User DirectedGraph Vertex Community PublishedContents
New Classes/Interfaces added:
1. GraphArchitect
2. IntroCSGraphBuilder
3. GraphBuilder

You can find the above tree in the GraphBuilders package. They are the implementation of builder design pattern that instantiate built-in Technical Trees.

4. Exceptions: a class that stores all the error messages we have so far.

Class I refactored:

1. Post & Community:
I changed the parameter of add() method so that it takes in only the content (in string) of the PublishedContents, instead of an instance of it. This reduces dependency.

2. CommunityLibrary, UserManager, GraphManager:
I added the instance variables, currentCommunity, currentUser to CommunityLibrary and UserManager respectively, and currentGraph and currentVertex to GraphManager. Each keeps track of an instance of the corresponding object that the program is working on.

3. commented out lines in Main, TempMainWalkThroughHelper, and some test files that would raise error due to the changes in Post and Comment.

New Test Class added:
SerializationTest: to lightly test whether storing a "CS Introductory Series" Technical Tree works or not. It is stored in Serialization.bin.