A commit of the full implementation of the achievement system. class involved:
Achievements (a constant class)
Userinfo

The mechanism of updating a user's achievement status is implemented in addPost() and incrementTotalLogins(). They use the checkAndRequestAchievement() method check whether the user is eligible for an achievement or not, if so, change the status of the corresponding achievement from false to true, in mapOfAchievement.

AchievementManager does not seem to be necessary. But it has been kept in case we need it in the future.