fixed Comment so that it obeys the parameters of its parent class.
fixed add() of Community and addPost() of CommunityLibrary so that they return the id of the post added.

add() of Post now also returns the id of the commented added.

changed the addPost() of UserInfo to addToListPostId(), as it now stores the id of all the post the corresponding User is adding.